requirements.txt
# Web Framework
fastapi==0.118.0
uvicorn==0.37.0

# Database
sqlalchemy==2.0.43
pymysql==1.1.2

# Data Validation & Settings
pydantic==2.11.9
pydantic-settings==2.11.0

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Environment & Configuration
python-dotenv==1.1.1

# Email Validation
email-validator==2.3.0

# Caching
redis==6.4.0

# HTTP Client
requests==2.32.5
httpx==0.27.0

# AWS Services
boto3==1.35.99
botocore==1.35.99

# Task Scheduling
apscheduler==3.10.4

# Email Services
# Note: smtplib is part of Python standard library, no need to install

# Additional dependencies based on code analysis
# uuid, json, datetime, enum, logging, random, hashlib, secrets, asyncio, typing, decimal, os
# These are Python standard library modules, no need to install separately